+++ src/main/java/svenhjol/charmony/ambient_sounds/client/features/biomes/sounds/Plains.java
@@ -10,8 +10,8 @@ import svenhjol.charmony.ambient_sounds.client.features.biomes.BiomeSound;
 import svenhjol.charmony.ambient_sounds.client.features.biomes.SurfaceBiomeSound;
 import svenhjol.charmony.ambient_sounds.client.features.sound.SoundHandler;
 import svenhjol.charmony.ambient_sounds.client.features.sound.SoundType;
-import svenhjol.charmony.ambient_sounds.helper.BiomeCheckHelper;
-import svenhjol.charmony.core.helper.WorldHelper;
+import svenhjol.charmony.ambient_sounds.helpers.BiomeCheckHelper;
+import svenhjol.charmony.core.helpers.WorldHelper;
 
 import javax.annotation.Nullable;