+++ src/main/java/svenhjol/charmony/ambient_sounds/client/features/environment/sounds/Mansion.java
@@ -12,7 +12,7 @@
  import svenhjol.charmony.ambient_sounds.client.features.environment.RepeatingEnvironmentSound;
  import svenhjol.charmony.ambient_sounds.client.features.sound.SoundHandler;
  import svenhjol.charmony.ambient_sounds.client.features.sound.SoundType;
- import svenhjol.charmony.core.helper.WorldHelper;
+ import svenhjol.charmony.core.helpers.WorldHelper;
 
  import java.util.List;