+++ src/main/java/svenhjol/charmony/ambient_sounds/client/features/environment/sounds/Snowstorm.java
@@ -8,8 +8,8 @@
  import svenhjol.charmony.ambient_sounds.client.features.environment.RepeatingEnvironmentSound;
  import svenhjol.charmony.ambient_sounds.client.features.sound.SoundHandler;
  import svenhjol.charmony.ambient_sounds.client.features.sound.SoundType;
- import svenhjol.charmony.ambient_sounds.helper.BiomeCheckHelper;
- import svenhjol.charmony.core.helper.WorldHelper;
+ import svenhjol.charmony.ambient_sounds.helpers.BiomeCheckHelper;
+ import svenhjol.charmony.core.helpers.WorldHelper;
 
  public class Snowstorm implements SoundType<EnvironmentSound> {
      public final SoundEvent sound;